export util function getIntolerances
